+++ mobile-app/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## mobile-app@v0.2.1+4
+
+- Disabled autocomplete, autocapitalize and spell check on typing textarea
+
 ## mobile-app@v0.2.0+3
 
 - Add preset feature